+++ src/Git_Updater/Base.php
@@ -517,7 +517,11 @@ class Base {
 		$new_source = $this->fix_misnamed_directory( $new_source, $remote_source, $upgrader_object, $slug );
 
 		if ( $source !== $new_source ) {
-			$this->move( $source, $new_source );
+			if ( function_exists( 'move_dir' ) ) {
+				move_dir( $source, $new_source );
+			} else {
+				$this->move_dir( $source, $new_source );
+			}
 		}
 
 		return trailingslashit( $new_source );