+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* use `move_dir()` and `is_virtualbox()` from [#51875](https://core.trac.wordpress.org/ticket/51857) [PR #2225](https://github.com/WordPress/wordpress-develop/pull/2225/files)
 
 #### 10.7.2 / 2022-03-31
 * fix directory rename for single file plugin update